Adding a check to see if api server is in profile or exec of k8s client in profile (#87)
