+++ chart/kubecop/crds/runtime-rule-binding.crd.yaml
@@ -97,12 +97,13 @@ spec:
                       - R0003
                       - R0004
                       - R0005
+                      - R0006
+                      - R0007
                       - R1000
                       - R1001
                       - R1002
                       - R1003
                       - R1004
-                      - R1005
                       - R1006
                       - R1007
                       type: string
@@ -114,11 +115,11 @@ spec:
                       - Unexpected capability used
                       - Unexpected domain request
                       - Unexpected Service Account Token Access
+                      - Kubernetes Client Executed
                       - Exec from malicious source
                       - Kernel Module Load
                       - Exec Binary Not In Base Image
                       - Malicious SSH Connection
-                      - Kubernetes Client Executed
                       - Exec from mount
                       - Unshare System Call usage
                       - Crypto Miners port detected
